Category 부분 수정 및 NoticeDetil Repository 수정 및 Service 수정 Controller 추가

+++ backend/src/main/java/com/example/demo/dao/CategoryService.java
@@ -18,8 +18,20 @@ public class CategoryService {
     }
 
 
-    public void DeleteCategory(Long categoryIdx){
-        categoryRepository.DeleteCategory(categoryIdx);
+    public CategoryVO DeleteCategory(Long categoryIdx){
+        CategoryVO findCategoryVO = null;
+        try{
+            //먼저 삭제할 것이 있는지 확인해주고
+            findCategoryVO = categoryRepository.findCategoryByIdx(categoryIdx).get(0);
+        }catch(NullPointerException e){
+            //없다면 null처리
+            findCategoryVO = null;
+        }finally {
+            if(findCategoryVO != null){
+                categoryRepository.DeleteCategory(categoryIdx);
+            }
+        }
+        return findCategoryVO;
     }
 
 
@@ -30,5 +42,9 @@ public class CategoryService {
 
     public List<CategoryVO> getCategoryList(Long memberIdx){
         return categoryRepository.GetCategoryList(memberIdx);
+    }
+
+    public List<CategoryVO> findCategoryByIdx(Long idx){
+        return categoryRepository.findCategoryByIdx(idx);
     }
 }
